+++ src/grammar/GrammarConstants.ts
@@ -22,8 +22,8 @@ GrammarConstants.group = "@group"
 GrammarConstants.any = "@any"
 
 // parser/vm instantiating and executing
-GrammarConstants.parser = "@parser"
-GrammarConstants.parserJs = "js"
+GrammarConstants.constructor = "@constructor"
+GrammarConstants.constructorJs = "js"
 
 // compiling
 GrammarConstants.compilerKeyword = "@compiler"